change hour count type from integer to double

+++ database/migrations/after/2022_08_05_122923_create_hours_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
     {
         Schema::create('hours', static function (Blueprint $table) {
             $table->id();
-            $table->integer('count');
+            $table->double('count');
             $table->date('date');
             $table->string('description')->nullable();
             $table->timestamps();
